Added peak calling and bedfile export functionality to ensemble analyzer. Also improved its memory footprint when dealing with many genes in a single input file.

+++ structure.cpp
@@ -4,7 +4,7 @@
 
 #include "structure.h"
 
-Loci::Loci(std::string C, std::string c, int s, int e): chromosome(C), strand(c), start_pos(s), end_pos(e) {}
+Loci::Loci(std::string C, std::string c, long int s, long int e): chromosome(C), strand(c), start_pos(s), end_pos(e) {}
 
 int Loci::get_length(){
     return end_pos - start_pos;
